requirements.txt
streamlit==1.35.0
anthropic==0.26.0
requests==2.31.0
beautifulsoup4==4.12.3
pdfplumber==0.11.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.4
